refactor: simplify redux import
Changes DataSlice file name to index so one less import level is required
