+++ frontend/src/modules/Workplace/information/FileTransferLabels/TransferLabelsDialog.jsx
@@ -40,8 +40,8 @@ import {
   uploadLabels,
   downloadLabels,
   downloadModel,
-} from "../../redux/DataSlice";
-import { curCategoryNameSelector } from "../../redux/DataSlice";
+} from "../../redux";
+import { curCategoryNameSelector } from "../../redux";
 import { blue } from "@mui/material/colors";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import ExpandLessIcon from "@mui/icons-material/ExpandLess";
