+++ frontend/src/modules/Workplace/redux/index.js
