+++ frontend/src/modules/Workplace/upperbar/Modal/CreateCategoryModal.jsx
@@ -20,7 +20,7 @@ import Typography from "@mui/material/Typography";
 import Modal from "@mui/material/Modal";
 import { useDispatch } from "react-redux";
 import { isFulfilled } from "@reduxjs/toolkit";
-import { createCategoryOnServer } from "../../redux/DataSlice";
+import { createCategoryOnServer } from "../../redux";
 import TextField from "@mui/material/TextField";
 import classes from "./index.module.css";
 import {
