+++ frontend/src/modules/Workplace/upperbar/Modal/DeleteCategoryModal.jsx
@@ -18,7 +18,7 @@ import Button from "@mui/material/Button";
 import { Dialog } from "@mui/material";
 import { DialogTitle, DialogActions, DialogContent, DialogContentText } from "@mui/material";
 import { useSelector, useDispatch } from "react-redux";
-import { curCategoryNameSelector, deleteCategory } from "../../redux/DataSlice";
+import { curCategoryNameSelector, deleteCategory } from "../../redux";
 import { notify } from "../../../../utils/notification";
 
 export default function DeleteCategoryModal({ open, setOpen }) {
