+++ frontend/src/store/configureStore.js
@@ -17,7 +17,7 @@ import { configureStore } from '@reduxjs/toolkit'
 import { workspacesReducer } from '../modules/Workspace-config/workspaceConfigSlice'
 import { authenticateReducer } from '../modules/Login/LoginSlice'
 import { featureFlagsReducer } from '../featureFlags/featureFlagsSlice'
-import workspaceReducer from '../modules/Workplace/redux/DataSlice'
+import workspaceReducer from '../modules/Workplace/redux'
 import { errorReducer } from '../error/errorSlice'
 import { errorMiddleware } from '../error/errorMiddleware'
